update config move sourcify files changed tests structure
