+++ verification/src/cli.rs
@@ -3,8 +3,6 @@ use clap::Parser;
 #[derive(Parser)]
 #[clap(author, version, about, long_about = None)]
 pub struct Args {
-    #[clap(short, long, default_value = "0.0.0.0")]
-    pub address: String,
-    #[clap(short, long, default_value = "8043")]
-    pub port: u16,
+    #[clap(short, long, default_value = "config.toml")]
+    pub config_path: std::path::PathBuf,
 }
