+++ verification/src/http_server/handlers/mod.rs
@@ -1,3 +1,10 @@
+use serde::Serialize;
+
 pub mod compilation;
 pub mod status;
 pub mod verification;
+
+#[derive(Debug, Serialize, PartialEq)]
+pub struct VerificationResponse {
+    pub verified: bool,
+}
