+++ verification/src/http_server/handlers/verification/mod.rs
@@ -1,2 +1,3 @@
 pub mod routes;
 pub mod solidity;
+pub mod sourcify;
