+++ verification/src/http_server/handlers/verification/solidity/flatten.rs
@@ -1,5 +1,8 @@
-use super::types::{FlattenedSource, VerificationRequest, VerificationResponse};
-use crate::{download_cache::DownloadCache, solidity::fetcher::SvmFetcher};
+use super::types::{FlattenedSource, VerificationRequest};
+use crate::{
+    download_cache::DownloadCache, http_server::handlers::VerificationResponse,
+    solidity::fetcher::SvmFetcher,
+};
 use actix_web::{
     error,
     web::{self, Json},
