+++ verification/src/http_server/handlers/verification/solidity/mod.rs
@@ -1,3 +1,2 @@
 pub mod flatten;
-pub mod sourcify;
 mod types;
