+++ verification/src/lib.rs
@@ -1,9 +1,9 @@
 mod cli;
-mod configuration;
+mod config;
 mod download_cache;
 mod http_server;
 mod solidity;
 
-pub use configuration::Config;
+pub use crate::config::Config;
 pub use http_server::routes;
 pub use http_server::server::run_server as run_http_server;
