--- verification/tests/endpoints/mod.rs
@@ -1,1 +1,0 @@
-mod sourcify;
