--- verification/tests/integration.rs
@@ -1,1 +1,0 @@
-mod endpoints;
